# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nihongo Learning API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/content/audio/{audio_id}": {
      "get": {
        "tags": [
          "Public Content"
        ],
        "summary": "Stream Audio",
        "operationId": "Public Content-stream_audio_api_v1_content_audio_audio_id",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/content/audio/upload": {
      "post": {
        "tags": [
          "Admin Content"
        ],
        "summary": "Upload Audio",
        "operationId": "Admin Content-upload_audio_api_v1_admin_content_audio_upload",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_Admin_Content-upload_audio_api_v1_admin_content_audio_upload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioAssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/content/import/json": {
      "post": {
        "tags": [
          "Admin Content"
        ],
        "summary": "Import Vocab Json",
        "operationId": "Admin Content-import_vocab_json_api_v1_admin_content_import_json",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/VocabularyCreate"
                },
                "type": "array",
                "title": "Vocabularies"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioAssetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          }
        },
        "type": "object",
        "required": [
          "id",
          "file_url",
          "duration_seconds"
        ],
        "title": "AudioAssetResponse"
      },
...
      "Body_Admin_Content-upload_audio_api_v1_admin_content_audio_upload": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "transcript": {
            "type": "string",
            "title": "Transcript"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_Admin Content-upload_audio_api_v1_admin_content_audio_upload"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VocabularyCreate": {
        "properties": {
          "word": {
            "type": "string",
            "title": "Word"
          },
          "kana": {
            "type": "string",
            "title": "Kana"
          },
          "romaji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Romaji"
          },
          "meaning": {
            "type": "string",
            "title": "Meaning"
          },
          "audio_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Id"
          }
        },
        "type": "object",
        "required": [
          "word",
          "kana",
          "meaning"
        ],
        "title": "VocabularyCreate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}